Phasing out mutable for immutable CashTransaction
- and refactored budget classes to support incremental updates

+++ src/main/java/bank2budget/core/BalanceHistory.java
@@ -1,6 +1,7 @@
 package bank2budget.core;
 
 import bank2budget.core.Account;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.Collections;
 import java.util.NavigableMap;
@@ -13,9 +14,9 @@ import java.util.TreeMap;
 public class BalanceHistory {
 
     private final Account account;
-    private final NavigableMap<LocalDate, Double> balances;
+    private final NavigableMap<LocalDate, BigDecimal> balances;
 
-    public BalanceHistory(Account account, NavigableMap<LocalDate, Double> balances) {
+    public BalanceHistory(Account account, NavigableMap<LocalDate, BigDecimal> balances) {
         this.account = account;
         this.balances =  Collections.unmodifiableNavigableMap(new TreeMap<>(balances));;
     }
@@ -28,7 +29,7 @@ public class BalanceHistory {
         return balances.lastKey();
     }
 
-    public Double getBalanceOn(LocalDate date) {
+    public BigDecimal getBalanceOn(LocalDate date) {
         return balances.get(date);
     }
